pyproject.toml
[project]
name = "python-playwright-saucedemo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pytest-playwright>=0.7.1",
]
